==> 3_building_frequencies.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import re
import string
import nltk 
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer
from nltk.tokenize import TweetTokenizer
import ast

df = pd.read_csv('Stemmed_Data2.csv',index_col=0)

df.head()

def build_freqs(tweets, ys):
    yslist = np.squeeze(ys).tolist()
    freqs_0 = {}
    freqs_1 = {}
    freqs_2 = {}
    for y, tweet in zip(yslist, tweets):
        for word in ast.literal_eval(tweet):
            if y==0:
                pair = word
                if pair in freqs_0:
                    freqs_0[pair] += 1
                else:
                    freqs_0[pair] = 1
            elif y==2:
                pair = word
                if pair in freqs_2:
                    freqs_2[pair] += 1
                else:
                    freqs_2[pair] = 1
            else:
                pair = word
                if pair in freqs_1:
                    freqs_1[pair] += 1
                else:
                    freqs_1[pair] = 1

    return freqs_0,freqs_1,freqs_2

freqs_0,freqs_1,freqs_2 = build_freqs(df['stemmed'].values, df['class'].values)

# print(freqs_0)

sort_freq_0 = sorted(freqs_0.items(), key=lambda x: x[1], reverse=True)
sort_freq_1 = sorted(freqs_1.items(), key=lambda x: x[1], reverse=True)
sort_freq_2 = sorted(freqs_2.items(), key=lambda x: x[1], reverse=True)

# print(sort_freq_0)

def remove(l,count):
    l = [v for i, v in enumerate(l) if v[1] >= count]
    return l

count = 5
filtered_0 = remove(sort_freq_0,count)
filtered_1 = remove(sort_freq_1,count)
filtered_2 = remove(sort_freq_2,count)

print(len(filtered_0))
print(len(filtered_1))
print(len(filtered_2))

def ordered_words(l):
    lst = [v[0] for i, v in enumerate(l) ]
    return lst

list_0 = ordered_words(filtered_0)
list_1 = ordered_words(filtered_1)
list_2 = ordered_words(filtered_2)

# print(list_0)

# if v[1]>=a.get(v[0],1) and v[1]>b.get(v[0],1)
def relative_freq(l,a,b):
    lst = [(v[0],v[1],a.get(v[0],1),b.get(v[0],1)) for i, v in enumerate(l)]
    return lst 
def relative_freq_filter(l,a,b):
    lst = [(v[0],v[1],a.get(v[0],1),b.get(v[0],1)) for i, v in enumerate(l) if v[1]>=a.get(v[0],1) or v[1]>b.get(v[0],1)]
    return lst 

relative_0 = relative_freq(filtered_0,freqs_1,freqs_2)
relative_1 = relative_freq(filtered_1,freqs_0,freqs_2)
relative_2 = relative_freq(filtered_2,freqs_0,freqs_1)
relative_filter_0 = relative_freq(filtered_0,freqs_1,freqs_2)
relative_filter_1 = relative_freq(filtered_1,freqs_0,freqs_2)
relative_filter_2 = relative_freq(filtered_2,freqs_0,freqs_1)

print(relative_filter_2)

def remove(l,count):
    l = [v for i, v in enumerate(l) if v[1] >= count]
    return l


lines1 = []
with open('/Users/femindharamshi/Downloads/MP 2/list1.txt') as f:
    lines1 = f.read().splitlines()

lines = []
with open('/Users/femindharamshi/Downloads/MP 2/list.txt') as f:
    lines = f.read().splitlines()

print(lines)

def relative_freq2(l,a,b,lines):
    lst = [(v[0],v[1],a.get(v[0],1),b.get(v[0],1)) for i, v in enumerate(l) if v[0] not in lines and v[1]>=40 ]
    return lst 

relative_1 = relative_freq2(filtered_1,freqs_0,freqs_2,lines)

relative_2_filter = relative_freq_filter(filtered_2,freqs_0,freqs_1)

list2_words = ordered_words(relative_1)

for i in list2_words:
    print(i)

len(list2_words)

combined = []
with open('/Users/femindharamshi/Downloads/MP 2/combined.txt') as f:
    combined = f.read().splitlines()

def relative_freq_combined(l,a,b,c):
    lst = [(v,a.get(v,1),b.get(v,1),c.get(v,1)) for v in l]
    return lst 
    
    

combined_freq = relative_freq_combined(combined, freqs_0, freqs_1, freqs_2)

print(combined_freq)
