==> 4_unigrams.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import re
import string
import nltk 
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer
from nltk.tokenize import TweetTokenizer
import ast

df = pd.read_csv('Stemmed_Data2.csv')
df.iloc[10]

cols = []
with open('C:/Users/femin/MP_Final/combined.txt') as f:
    cols = f.read().splitlines()

len(cols)


print('hoe' in cols)

def generate_col(row,col):
    if col in ast.literal_eval(row):
        return 1
    else:
        return 0

def count_col(row):
    count = 0
    for i in ast.literal_eval(row):
        if i in cols:
            count=count+1
    return count

for col in cols:
    df[col]=df['stemmed'].apply(lambda x: generate_col(x,col))
# df['count']=df['stemmed'].apply(count_col)

df.head()

print(list(df.columns))

df.to_csv('unigram3.csv')

df1= df.groupby('faggot')['faggot'].count()
print(df1)

# count_col(['dawg', 'ever', 'fuck', 'bitch', 'start'] )

print(df['tweet'].iloc[236:237].values)
