==> 5_Sentiment Analysis.ipynb <==
from sentistrength import PySentiStr
import pandas as pd

senti = PySentiStr()
senti.setSentiStrengthPath('C:/Users/femin/MP_Final/SentiStrength.jar') # Note: Provide absolute path instead of relative path
senti.setSentiStrengthLanguageFolderPath('C:/Users/femin/MP_Final/SentiStrength/') # Note: Provide absolute path instead of relative path

data = pd.read_csv('Stemmed_Data2.csv')
data.tail()


df1 = data['tweet']

df1.head()

arr = df1.to_numpy()

print(arr[-5])

scale = senti.getSentiment(arr, score = 'scale')
dual = senti.getSentiment(arr, score = 'dual')
binary = senti.getSentiment(arr, score = 'binary')


data[['Pos','Neg']] =  pd.DataFrame(dual,columns=['Pos','Neg'])


data.tail()



data["scale"]= pd.DataFrame(scale)
data["binary"]= pd.DataFrame(binary)
data.tail()

# Dnew = pd.concat([data, data['scale'], data['dual'], data['bin']], axis=1)


# Dnew.head()


# data = data.drop('binary',axis=1)
# data["binary"]= pd.DataFrame(binary)
data.head()

data.shape

data.to_csv('sentistrength.csv')

def getScale(var):
    score = senti.getSentiment(var, score = 'scale')
    return score

data["scale"]= data['tweet'].apply(getScale)

tweet = "She is total bitch, She should die"

print("Postive,Negative:",senti.getSentiment(tweet, score = 'dual'))
print("Overall:",senti.getSentiment(tweet, score = 'scale'))
print("Binary:",senti.getSentiment(tweet, score = 'binary'))
