==> 6_combine_datasets.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import re
import string
import nltk 
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer
from nltk.tokenize import TweetTokenizer
import ast

sem_sem = pd.read_csv('stemmed_semantic_data2.csv',index_col=0)

sem_sem.tail()

uni = pd.read_csv('unigram3.csv')
uni.head()

senti = pd.read_csv('sentistrength.csv',index_col=0)
senti.tail()

resultant = pd.read_csv('Stemmed_Data2.csv')

resultant.head()

resultant = resultant.drop(['tweet','processed_text','stemmed'],axis = 1)
resultant.head()

sem_sem.columns

resultant[['scale', 'Pos', 'Neg', 'binary']] = senti[[ 'scale', 'Pos', 'Neg', 'binary']]

resultant[['QuestionMarkCount', 'QuotationMarkCount', 'ExclamationMarkCount',
       'AllCapitalised']] = sem_sem[['QuestionMarkCount', 'QuotationMarkCount', 'ExclamationMarkCount',
       'AllCapitalised']]

resultant = resultant.loc[:, ~resultant.columns.str.contains('^Unnamed')]
resultant.tail()

cols = []
with open('C:/Users/femin/MP_Final/combined.txt') as f:
    cols = f.read().splitlines()

for i in cols:
    resultant[i] = uni[i]
# resultant['count'] = uni['count']

resultant.tail()

resultant.to_csv('final_dataset3.csv')

uni.tail()

uni.shape

sem_sem.tail()

resultant.iloc
