==> 7_Classifier.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import re
import string
import scipy as sp
import seaborn as sns
import time
%matplotlib inline

import pickle

from sklearn import preprocessing
from sklearn.preprocessing import MinMaxScaler,StandardScaler
scaler = StandardScaler()
minmax = MinMaxScaler()

data = pd.read_csv('final_dataset3.csv',index_col=0)
# data2 = pd.read_csv('final_dataset2.csv',index_col=0)
# data['count'] = data2['count']
# data['class'] = data['class'].replace([0],1)

data.head(100)

# data = data.drop(['Pos','Neg','binary'],axis=1)
columns = data.columns

print(columns)

print(data.head())
data.describe

X = data[columns[1:]].to_numpy()
X = minmax.fit_transform(X)
y = data['class'].to_numpy()

print(X)

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test=train_test_split(X,y,test_size=0.2,stratify=data['class'])

from sklearn.ensemble import RandomForestClassifier
clf = RandomForestClassifier(max_depth=40, random_state=0)

%time clf.fit(X_train, y_train)

%time Score = clf.score(X_test,y_test)
print(Score*100)

from sklearn.metrics import accuracy_score, confusion_matrix, classification_report, cohen_kappa_score

%time y_pred = clf.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print('Accuracy', accuracy_score(y_test, y_pred) * 100)
print(classification_report(y_test, y_pred))
print('Cohen Kappa score:', cohen_kappa_score(y_test, y_pred))

print(y_pred)

from sklearn.linear_model import LogisticRegression
lr=LogisticRegression()
st = time.time()
lr.fit(X_train,y_train)
et = time.time()
print(et-st)
%time y_pred=lr.predict(X_train)
print(confusion_matrix(y_train, y_pred))
print('Accuracy', accuracy_score(y_train, y_pred) * 100)
print(classification_report(y_train, y_pred))
print('Cohen Kappa score:', cohen_kappa_score(y_train, y_pred))

X.shape

Pkl_Filename = "ternary_model.pkl"  
with open(Pkl_Filename, 'wb') as file:  
    pickle.dump(lr, file)
  


from sklearn.neighbors import KNeighborsClassifier
knn = KNeighborsClassifier(n_neighbors=10)
%time knn.fit(X_train,y_train)

%time y_pred = knn.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print('Accuracy', accuracy_score(y_test, y_pred) * 100)
print(classification_report(y_test, y_pred))
print('Cohen Kappa score:', cohen_kappa_score(y_test, y_pred))

from sklearn.naive_bayes import GaussianNB

gnb = GaussianNB()
%time model = gnb.fit(X_train, y_train)

%time y_pred = model.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print('Accuracy', accuracy_score(y_test, y_pred) * 100)

print(classification_report(y_test, y_pred))
print('Cohen Kappa score:', cohen_kappa_score(y_test, y_pred))

from sklearn.metrics import accuracy_score, log_loss
from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import SVC, LinearSVC, NuSVC
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier, GradientBoostingClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis


classifiers = [
#     KNeighborsClassifier(3),
#     SVC(kernel="rbf", C=0.025, probability=True),
#     NuSVC(probability=True),
    DecisionTreeClassifier(),
    RandomForestClassifier(),
    AdaBoostClassifier(),
    GradientBoostingClassifier(),
    GaussianNB(),
    LinearDiscriminantAnalysis(),
    QuadraticDiscriminantAnalysis()]

log_cols=["Classifier", "Accuracy", "Log Loss"]
log = pd.DataFrame(columns=log_cols)

for clf in classifiers:
    clf.fit(X_train, y_train)
    name = clf.__class__.__name__
    
    print("="*30)
    print(name)
    
    print('****Results****')
    train_predictions = clf.predict(X_test)
    acc = accuracy_score(y_test, train_predictions)
    print("Accuracy: {:.4%}".format(acc))
    
    train_predictions = clf.predict_proba(X_test)
    ll = log_loss(y_test, train_predictions)
    print("Log Loss: {}".format(ll))
    
    log_entry = pd.DataFrame([[name, acc*100, ll]], columns=log_cols)
    log = log.append(log_entry)
    
print("="*30)
