==> 10_DeepLearningBinary.ipynb <==
import tensorflow as tf
from sklearn import preprocessing
from sklearn.preprocessing import MinMaxScaler,StandardScaler
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import re
import string
import scipy as sp
import seaborn as sns
import time
%matplotlib inline

scaler = StandardScaler()
minmax = MinMaxScaler()

data = pd.read_csv('final_dataset3.csv',index_col=0)
# data['class'] = data['class'].replace([1],0)
# data['class'] = data['class'].replace([2],1)

data.shape

columns = data.columns
X = data[columns[1:]].to_numpy()
X = tf.keras.utils.normalize(X, axis=1)
y = data['class'].to_numpy()

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test=train_test_split(X,y,test_size=0.2,stratify=data['class'])

model = tf.keras.models.Sequential()
model.add(tf.keras.layers.Dense(128, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(256, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(512, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(2, activation=tf.nn.softmax))

model.compile(optimizer='adam',
              loss='sparse_categorical_crossentropy',
              metrics=['accuracy'])

model.fit(X_train, y_train, epochs=150, batch_size=70)

print("\nAccuracy on Test Data: \n")

val_loss, val_acc = model.evaluate(X_test,y_test)

model = tf.keras.models.Sequential()
model.add(tf.keras.layers.Dense(256, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(128, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(256, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(512, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(2, activation=tf.nn.sigmoid))

model.compile(optimizer='adam',
              loss='sparse_categorical_crossentropy',
              metrics=['accuracy'])

model.fit(X_train, y_train, epochs=150, batch_size=70)
print("\nAccuracy on Test Data: \n")
val_loss, val_acc = model.evaluate(X_test,y_test)

model = tf.keras.models.Sequential()
model.add(tf.keras.layers.Dense(1024, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(512, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(256, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(128, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(2, activation=tf.nn.sigmoid))

model.compile(optimizer='adam',
              loss='sparse_categorical_crossentropy',
              metrics=['accuracy'])

model.fit(X_train, y_train, epochs=150, batch_size=70)
print("\nAccuracy on Test Data: \n")
val_loss, val_acc = model.evaluate(X_test,y_test)

model = tf.keras.models.Sequential()
model.add(tf.keras.layers.Dense(256, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(128, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(256, activation=tf.nn.relu))
model.add(tf.keras.layers.Dense(2, activation=tf.nn.sigmoid))

model.compile(optimizer='adam',
              loss='sparse_categorical_crossentropy',
              metrics=['accuracy'])

model.fit(X_train, y_train, epochs=1000, batch_size=70)
print("\nAccuracy on Test Data: \n")
val_loss, val_acc = model.evaluate(X_test,y_test)
